pyproject.toml
[tool.poetry]
name = "clinvar-data-monitor"
version = "0.2.0"
description = "Automated data quality monitoring for ClinVar with Quilt versioning"
authors = ["Lina Faller <lina.faller@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8.1"
quilt3 = ">=5.0"
pandas = "^2.0.0"
requests = "^2.31.0"
boto3 = "^1.26.0"
pyyaml = "^6.0"
plotly = "^5.16.0"
python-dateutil = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"
ipython = "^8.0.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=html"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
